# api/openapi.yaml
openapi: 3.0.3
info:
  title: ОбъектМаркет API
  version: 1.0.0
  description: 'API учебной системы закупки строительных материалов. Суммы — в копейках; реальные платежи и GPS отсутствуют. Демонстрационная доставка доступна только для строк адреса вида «Астрахань, улица, дом» или «г. Астрахань, улица, дом».'
servers:
  - url: /api/v1
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: om_session
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
...
    CartItem:
      type: object
      required: [offerId, quantity]
      properties:
        offerId: {type: string, format: uuid}
        quantity: {type: integer, minimum: 1, maximum: 10000}
    Purchase:
      type: object
      required: [items, address]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items: {$ref: '#/components/schemas/CartItem'}
        address: {type: string, description: 'Учебный адрес: «Астрахань, улица, дом» или «г. Астрахань, улица, дом»'}
        requestedDate: {type: string, format: date}
        projectId: {type: string, format: uuid}
...
    Order:
      type: object
      properties:
        id: {type: string, format: uuid}
        status: {type: string, enum: [awaiting_payment, paid, in_progress, delivered, cancelled, expired]}
        reservationExpiresAt: {type: string, format: date-time, nullable: true, description: 'Через 30 минут неоплаченный резерв освобождается'}
        itemsTotalKopecks: {type: integer}
        deliveryTotalKopecks: {type: integer}
        totalKopecks: {type: integer}
        items: {type: array, items: {type: object}}
        deliveries: {type: array, items: {type: object}}
        paymentMode: {type: string, enum: [demo]}
        trackingMode: {type: string, enum: [demo]}
...
  responses:
    Error:
      description: Ошибка с кодом и сообщением
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
paths:
...
  /quotes:
    post:
      summary: Рассчитать стоимость без резервирования
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/Purchase'}}}
      responses: {'201': {description: 'Строки, итоги, недоступные позиции, предупреждения и zoneAvailable для учебной зоны'}}
...
  /orders:
    get:
      summary: Заказы покупателя
      security: [{session: []}]
      responses: {'200': {description: История заказов}}
    post:
      summary: Создать заказ и транзакционно зарезервировать остатки
      security: [{session: []}]
      parameters: [{name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 100, pattern: '^[A-Za-z0-9._:-]+$'}, description: 'Повтор с тем же ключом и телом возвращает тот же заказ; другое тело — 409'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/Purchase'}}}
      responses: {'201': {description: Заказ и поставки по поставщикам, резерв на 30 минут}, '400': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
  /orders/{id}:
    get:
      summary: Собственный заказ
      security: [{session: []}]
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Заказ, позиции и поставки, content: {application/json: {schema: {$ref: '#/components/schemas/Order'}}}}
        '404': {$ref: '#/components/responses/Error'}
  /orders/{id}/demo-payment:
    post:
      summary: Учебное подтверждение оплаты без списания средств
      security: [{session: []}]
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'201': {description: Обновлённый заказ}}
  /orders/{id}/cancel:
    post:
      summary: Отменить неоплаченный заказ и восстановить остатки
      description: Повторная отмена безопасна. Оплаченный заказ отменить нельзя. Истёкший резерв получает статус expired.
      security: [{session: []}]
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'201': {description: Обновлённый заказ}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
...
  /geo/route:
    get:
      summary: Маршрут 2ГИС или схематический маршрут демо
      parameters:
        - {name: fromLon, in: query, required: true, schema: {type: number}}
        - {name: fromLat, in: query, required: true, schema: {type: number}}
        - {name: toLon, in: query, required: true, schema: {type: number}}
        - {name: toLat, in: query, required: true, schema: {type: number}}
      responses: {'200': {description: 'Источник, координаты [долгота, широта], дистанция и время'}, '502': {$ref: '#/components/responses/Error'}}